Create Week4_A2_Mini_Web_Crawler
This script accepts a URL input from the user, parses it with Beautiful Soup, goes to the nth link entered by the user (position) and continues to crawl sub-links further for nth time (counts) which is also entered by the user. For the sample data provided by Coursera, using RegEx, it extracts the name displayed in the last list.